refactor:change KsqlAuthorizationException to KsqlTopicAuthorizationException
